Update of project
