+++ src/test/resources/features/Test_Troubleshoot.feature
@@ -1,0 +1,6 @@
+@Troubleshoot
+Feature:
+
+  Scenario:
+    Given I am on the login page
+    And I login as a librarian
