+++ src/test/resources/features/US04_BZ.feature
@@ -3,16 +3,16 @@ Feature: As a user, I want to select different number of rows to see users recor
 
 
 Scenario Outline: Show records for <count> options
-Given I am on the login page
+ Given I am on the login page
  And I login as a librarian
  And I click on "Users" link
- #When I select Show <count> records
- #Then show records value should be <count>
- #And the users table must display <count> records
-Examples:
- | count |
- | 5 |
- | 10 |
- | 15 |
- | 50 |
- | 100 |
+ When I select Show <count> records
+ Then show records value should be <count>
+ And the users table must display <count> records
+ Examples:
+  | count |
+  | 5 |
+  | 10 |
+  | 15 |
+  | 50 |
+  | 100 |
